requirements.txt
kafka-python
elasticsearch
prometheus_client